+++ ribbon/src/main/java/feign/ribbon/LBClient.java
@@ -26,8 +26,10 @@ import com.netflix.loadbalancer.ILoadBalancer;
 
 import java.io.IOException;
 import java.net.URI;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.LinkedHashSet;
 import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Set;
@@ -36,6 +38,7 @@ import feign.Client;
 import feign.Request;
 import feign.RequestTemplate;
 import feign.Response;
+import feign.Util;
 
 public final class LBClient extends
     AbstractLoadBalancerAwareClient<LBClient.RibbonRequest, LBClient.RibbonResponse> {
@@ -113,12 +116,14 @@ public final class LBClient extends
     }
 
     Request toRequest() {
-      return new RequestTemplate()
-          .method(request.method())
-          .append(getUri().toASCIIString())
-          .headers(request.headers())
-          .body(request.body(), request.charset())
-          .request();
+      // add header "Content-Length" according to the request body
+      final byte[] body = request.body();
+      final int bodyLength = body != null ? body.length : 0;
+      // create a new Map to avoid side effect, not to change the old headers
+      Map<String, Collection<String>> headers = new LinkedHashMap<String, Collection<String>>();
+      headers.putAll(request.headers());
+      headers.put(Util.CONTENT_LENGTH, Arrays.asList(String.valueOf(bodyLength)));
+      return Request.create(request.method(), getUri().toASCIIString(), headers, body, request.charset());
     }
 
     Client client() {